+++ R/quadmesh.R
@@ -9,17 +9,9 @@
 
 #' @importFrom raster xmin xmax ymin ymax
 edgesXY <- function(x) {
-  ## report to Hijmans 2015-11-06
-  #extract(r, expand.grid(c(xmin(r), xmax(r)), c(ymin(r), ymax(r))), method = "bilinear")
-  #[1]   NA   NA 99.5   NA
-  ## remove this eps fudge once bilinear works
-  eps <- rep(sqrt(.Machine$double.eps), 2L)
   xx <- seq(xmin(x), xmax(x), length = ncol(x) + 1L)
   yy <- seq(ymax(x), ymin(x), length = nrow(x) + 1L)
-  xx[c(1L, length(xx))] <- xx[c(1L, length(xx))] + c(+eps[2], -eps[2])
-  yy[c(1L, length(yy))] <- yy[c(1L, length(yy))] + c(-eps[1], +eps[1])
-  xy <- expand.grid(x = xx,
-                        y = yy)
+  xy <- expand.grid(x = xx, y = yy)
   xy
 }
 
@@ -28,38 +20,64 @@ prs <- function(x) {
   cbind(head(x, -1), tail(x, -1))
 }
 
-p4 <- function(xp, nc) {
-  (xp + c(0, 0, rep(nc, 2)))[c(1, 2, 4, 3)]
-}
+## what was this for ...
+# p4 <- function(xp, nc) {
+#   (xp + c(0, 0, rep(nc, 2)))[c(1, 2, 4, 3)]
+# }
+
 
 
 #' Create a quad-type mesh for use in rgl.
 #'
-#' Convert an object to a \code{\link[rgl]{mesh3d}} quadrangle mesh,
-#' currently the only available method is for \code{\link[raster]{raster}}.
+#' Convert an object to a `mesh3d` ([rgl::qmesh3d()]) quadrangle mesh,
+#' with methods for [raster::raster()] and `matrix`.
+#'
+#' `quadmesh()` generates the cell-based interpretation of a raster (AREA) but applies a continuous
+#' interpretation of the values of the cells to each quad corner. `dquadmesh` splits the mesh and
+#' applies a discrete interpretation directly. Loosely, the quadmesh is a continuous surface and the dquadmesh
+#' is free-floating cells, but it's a little more complicated and depends on the options applied. (The interpolation)
+#' applied in the quadmesh case is not entirely consistent.
 #'
 #' The output is described as a mesh because it is a dense representation
 #' of a continuous shape, in this case plane-filling quadrilaterals defined
 #' by index of four of the available vertices.
 #'
+#' The `z` argument defaults to the input `x` argument, though may be set to `NULL`, a constant
+#' numeric value, or another raster. If the coordinate system of `z` and `x` don't match the z values
+#' are queried by reprojection.
+#'
 #' Any raster RGB object (3-layers, ranging in 0-255) may be used as
 #' a _texture_ on the resulting mesh3d object.
-#' It is not possible to provide rgl with an object of data for texture, it must be a PNG file.
+#' It is not possible to provide rgl with an object of data for texture, it must be a PNG file and so
+#' the in-memory `texture` argument is written out to PNG file (with a message). The location of the file
+#' may be set explicitly with `texture_filename`.  Currently it's not possible to not use the `texture` object
+#' in-memory.
+#'
+#'
 #' @param x raster object for mesh structure
 #' @param z raster object for height values
 #' @param na.rm remove quads where missing values?
 #' @param ... ignored
 #' @param texture optional input RGB raster, 3-layers
+#' @param texture_filename optional input file path for PNG texture
 #' @return mesh3d
 #' @export
+#' @aliases dquadmesh
 #' @importFrom raster extract extent values
 #' @importFrom png writePNG
+#' @importFrom sp SpatialPoints CRS
 #' @examples
 #' library(raster)
 #' data(volcano)
 #' r <- setExtent(raster(volcano), extent(0, 100, 0, 200))
 #' qm <- quadmesh(r)
-quadmesh <- function(x, z = x, na.rm = FALSE, ..., texture = NULL) {
+quadmesh <- function(x, z = x, na.rm = FALSE, ..., texture = NULL, texture_filename = NULL) {
+  UseMethod("quadmesh")
+}
+
+#' @name quadmesh
+#' @export
+quadmesh.BasicRaster <- function(x, z = x, na.rm = FALSE, ..., texture = NULL, texture_filename = NULL) {
   x <- x[[1]]  ## just the oneth raster for now
   exy <- edgesXY(x)
  # ind <- apply(prs(seq(ncol(x) + 1)), 1, p4, nc = ncol(x) + 1)
@@ -76,7 +94,21 @@ quadmesh <- function(x, z = x, na.rm = FALSE, ..., texture = NULL) {
     ind1 <- ind1[,!is.na(values(x))]
   }
   ob <- .mkq3d()
-  if (!is.null(z)) z <- zapsmall(extract(z, exy, method = "bilinear")) else z <- 0
+
+  if (!is.null(z)) {
+    ## wish of https://github.com/hypertidy/quadmesh/issues/17
+    sp_exy <- sp::SpatialPoints(exy, proj4string = sp::CRS(raster::projection(x), doCheckCRSArgs = FALSE))
+    old <- options(warn = -1)
+
+    ## rather than add eps to the edgesXY coordinates
+    ## let's extend the z raster a tiny bit
+    z <- raster::setExtent(z, raster::extent(z) + raster::res(z)/1000)
+
+    z <- zapsmall(extract(z, sp_exy, method = "bilinear"))
+    options(old)
+  } else {
+    z <- 0
+  }
   ob$vb <- t(cbind(exy, z, 1))
   ob$ib <- ind1
 
@@ -96,37 +128,41 @@ quadmesh <- function(x, z = x, na.rm = FALSE, ..., texture = NULL) {
   texcoords <- texture_coordinates(texture, vertices = exy)
 
    ob$texcoords <- t(texcoords)
-   pngfilename <- tempfile(fileext = ".png")
-   message(sprintf("writing texture image to %s", pngfilename))
-   png::writePNG(raster::as.array(texture) / 255, pngfilename)
-   ob$material$texture <- pngfilename
+   if (is.null(texture_filename)) {
+     texture_filename <- tempfile(fileext = ".png")
+   } else {
+     if (!grepl("png$", texture_filename)) {
+       warning(sprintf("'texture filename' does not look like a good PNG filename'%s'",
+                       texture_filename))
+     }
+   }
+      message(sprintf("writing texture image to %s", texture_filename))
+   png::writePNG(raster::as.array(texture) / 255, texture_filename)
+   ob$material$texture <- texture_filename
    ob$material$col <- "grey"
   }
+  rproj <- raster::projection(x)
+  ## chuck on the original structure md
+  ob$raster_metadata <- list(xmn = raster::xmin(x),
+                             xmx = raster::xmax(x),
+                             ymn = raster::ymin(x),
+                             ymx = raster::ymax(x),
+                             ncols = raster::ncol(x),
+                             nrows = raster::nrow(x),
+                             crs = rproj)
+  ob$crs <- rproj
+  class(ob) <- c("quadmesh", "mesh3d", "shape3d")
   ob
 }
 
-
-
-
-# quad <- function(x, z = x, na.rm = FALSE) {
-#   x <- x[[1]]  ## just the oneth raster for now
-#   exy <- edgesXY(x)
-#   ind <- apply(prs(seq(ncol(x) + 1)), 1, p4, nc = ncol(x) + 1)
-#   ## all face indexes
-#   ind0 <- as.vector(ind) +
-#     rep(seq(0, length = nrow(x), by = ncol(x) + 1), each = 4 * ncol(x))
-#   ind1 <- matrix(ind0, nrow = 4)
-#   ## need to consider normalizing vertices here
-#   if (na.rm) {
-#     ind1 <- ind1[,!is.na(values(x))]
-#   }
-#   if (!is.null(z)) z <- extract(z, exy, method = "bilinear") else z <- 0
-#   v <- data.frame(x_ = exy[,1], y_ = exy[,2], vertex_ = seq(nrow(exy)))
-#   print(dim(ind1))
-#   b <- data.frame(vertex_ = as.vector(ind1), quad_ = rep(seq(ncol(ind1)), each = nrow(ind1)))
-#   list(v = v, b = b)
-# }
-#
-
-# distinct ----------------------------------------------------------------
+#' @name quadmesh
+#' @export
+quadmesh.matrix <- function(x, z = x, na.rm = FALSE, ..., texture = NULL, texture_filename = NULL) {
+  x <- raster::setExtent(raster::raster(x), raster::extent(0, ncol(x), 0, nrow(x)))
+  if (is.matrix(z)) {
+    #warning("z is a matrix ...")
+    z <- x
+  }
+  quadmesh(x, z = z, na.rm = na.rm, ..., texture = texture, texture_filename = texture_filename)
+}